Tamiflex handling for Call Graph
Former-commit-id: d645559534a95b547993b42925697fd3c1348ec2

+++ OPAL/tac/src/main/scala/org/opalj/tac/fpcf/analyses/cg/reflection/ReflectionRelatedCallsAnalysis.scala
@@ -346,7 +346,7 @@ class MethodInvokeAnalysis private[analyses] (
 
     override def processNewCaller(
         caller:          DefinedMethod,
-        pc:              UShort,
+        pc:              Int,
         tac:             TACode[TACMethodParameter, V],
         receiverOption:  Option[Expr[V]],
         params:          Seq[Option[Expr[V]]],
